requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.5
colored==1.4.3
decorator==5.1.1
docopt==0.6.2
EbookLib==0.17.1
executing==0.8.3
halo==0.0.31
jedi==0.18.1
log-symbols==0.0.14
lxml==4.9.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
six==1.16.0
spinners==0.0.24
stack-data==0.3.0
termcolor==1.1.0
toml==0.10.2
traitlets==5.3.0
wcwidth==0.2.5
